Connect User&Video(Model), Add File Upload

+++ src/controllers/videoController.js
@@ -1,9 +1,10 @@
-import { redirect } from "express/lib/response";
 import Video from "../models/Video";
+import User from "../models/User";
 
 // In Home Router
 const handleHome = async (req, res) => {
   const videos = await Video.find({}).sort({ createdAt: "desc" });
+  console.log(videos);
   return res.render("videos/home", {
     pageTitle: "C-Tube",
     videos,
@@ -30,7 +31,8 @@ const handleSearch = async (req, res) => {
 // In Video Rotuer
 const handleWatch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("publisher");
+  // NOT Found Video(Model)
   if (!video) {
     return res.status(404).render("404", {
       pageTitle: "Video not Found.",
@@ -43,11 +45,17 @@ const handleWatch = async (req, res) => {
 };
 const handleGetEdit = async (req, res) => {
   const { id } = req.params;
+  const { _id } = req.session.user;
   const video = await Video.findById(id);
+  // NOT Found Edit Video
   if (!video) {
     return res.status(404).render("404", {
       pageTitle: "Video not Found.",
     });
+  }
+  // Edit Video Publisher ID !== Login User ID
+  if (String(video.publisher) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("videos/edit", {
     pageTitle: `Edit ${video.title}`,
@@ -56,22 +64,29 @@ const handleGetEdit = async (req, res) => {
 };
 const handlePostEdit = async (req, res) => {
   const { id } = req.params;
+  const { _id } = req.session.user;
   const { title, description, hashtags, category } = req.body; // form으로부터 받아온 값
-  const video = await Video.exists({ _id: id }); // exits()는 filter를 argument로 받음
+  // const video = await Video.exists({ _id: id }); // exits()는 filter를 argument로 받음
+  const video = await Video.findById(id);
+  // NOT Found Edit Video
   if (!video) {
     return res.render("404", {
       pageTitle: "Video not Found.",
     });
   }
-  await Video.findByIdAndUpdate(id, {
+  // Edit Video Publisher ID !== Login User ID
+  if (String(video.publisher) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  const updateVideo = await Video.findByIdAndUpdate(id, {
     title,
     description,
     category,
     hashtags: Video.formatHashtags(hashtags),
   });
+  console.log("Update Video!");
   return res.redirect(`/videos/${id}`);
 };
-
 const handleGetUpload = (req, res) => {
   return res.render("videos/upload", {
     pageTitle: "Upload Video",
@@ -79,16 +94,29 @@ const handleGetUpload = (req, res) => {
 };
 const handlePostUpload = async (req, res) => {
   // POST Request Form Data
-  const { title, description, hashtags, category } = req.body;
+  const {
+    session: {
+      user: { _id },
+    },
+    body: { title, description, hashtags, category },
+    file: { path },
+  } = req;
 
-  // create Video
+  // Create New Video
   try {
-    await Video.create({
+    const newVideo = await Video.create({
+      path,
       title,
       description,
       category,
       hashtags: Video.formatHashtags(hashtags),
+      publisher: _id,
     });
+    // Add New Video ID -> User's Videos Array
+    const user = await User.findById(_id);
+    user.myVideos.push(newVideo._id);
+    user.save();
+    return res.redirect("/");
   } catch (error) {
     console.log(error);
     return res.status(400).render("videos/upload", {
@@ -96,12 +124,26 @@ const handlePostUpload = async (req, res) => {
       errorMsg: true,
     });
   }
-  return res.redirect("/");
 };
-
-const handleDel = async (req, res) => {
+const handleDelVideo = async (req, res) => {
   const { id } = req.params;
+  const { _id } = req.session.user;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  // NOT Found Delete Video
+  if (!video) {
+    return res.status(404).render("404", {
+      pageTitle: "Video is Not Found",
+    });
+  }
+  // Delete Video Publisher ID !== Login User ID
+  if (String(video.publisher) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  // Delete Video & Delete User's myVideos Array
   await Video.findByIdAndDelete(id);
+  user.myVideos.splice(user.myVideos.indexOf(id), 1);
+  user.save();
   return res.redirect("/");
 };
 
@@ -111,7 +153,7 @@ export {
   handleWatch,
   handleGetEdit,
   handlePostEdit,
-  handleDel,
+  handleDelVideo,
   handleGetUpload,
   handlePostUpload,
 };
